Consider 'custom_read/write' in nully concepts. Issue:
Types that conform to the nully concepts and those for
standard read_json/write_json match multiple overloads:
1. the read_json/write_json for the nully
2. the read_json/write_json for the standard type

Desirable to allow nullable types to not only be wrappers around
objects, but also the objects themself. For example, json_t is
itself readable & writable, but also nullable w/o a wrapper.